personal_calendar_agents/requirements.txt
genai
python-dotenv
google-auth==2.40.1
google-genai==1.12.1
google-api-python-client
fastapi
uvicorn